refactor: combine functions into one file
since these functions like GetDevID, GetNodeIP ..., they all related
to the host system info, just move them into one file, and easy to
controll.
